% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gretaGP}
\alias{gretaGP}
\alias{gretaGP-package}
\title{Gaussian process modelling in greta}
\description{
A greta module to create and combine covariance functions and
  use them to build Gaussian process models in greta. See
  \code{\link{kernels}} and \code{\link{gp}}
}
