% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructors.R
\name{kernels}
\alias{kernels}
\alias{bias}
\alias{white}
\alias{linear}
\alias{rbf}
\alias{expo}
\alias{mat12}
\alias{mat32}
\alias{mat52}
\alias{periodic}
\title{Gaussian process kernels}
\usage{
bias(variance, dim = 1)

white(variance, dim = 1)

linear(variances, columns = seq_along(variances))

rbf(lengthscales, variance, columns = seq_along(lengthscales))

expo(lengthscales, variance, columns = seq_along(lengthscales))

mat12(lengthscales, variance, columns = seq_along(lengthscales))

mat32(lengthscales, variance, columns = seq_along(lengthscales))

mat52(lengthscales, variance, columns = seq_along(lengthscales))

periodic(period, lengthscale, variance, dim = 1, columns = seq_len(dim))
}
\arguments{
\item{variance, variances}{(scalar/vector) the variance of a Gaussian process
prior in all dimensions (\code{variance}) or in each dimension
(\code{variances})}

\item{dim}{(scalar integer, not a greta array) the dimension of the Gaussian
process (number of columns on which it acts)}

\item{columns}{(scalar/vector integer, not a greta array) the columns of the
data matrix on which this kernel acts. Must have the same dimensions as lengthscale
parameters.}

\item{period}{(scalar) the period of the Gaussian process}

\item{lengthscale, lengthscales}{(scalar/vector) the correlation decay
distance along all dimensions (\code{lengthscale}) or each dimension
((\code{lengthscales})) of the Gaussian process}
}
\description{
Create and combine Gaussian process kernels (covariance
  functions) for use in Gaussian process models.
}
\details{
The kernel constructor functions each return a \emph{function} (of
  class \code{greta_kernel}) which can be executed on greta arrays to compute
  the covariance matrix between points in the space of the Gaussian process.
  The \code{+} and \code{*} operators can be used to combine kernel functions
  to create new kernel functions.

  The kernels are imported from the GPflow python package, using the gpflowr
  R package. Both of those need to be installed before you can use these
  methods. See the \href{gpflow.readthedocs.io}{GPflow website} for details
  of the kernels implemented.
}
\examples{
# create a radial basis function kernel on two dimensions
k1 <- rbf(lengthscales = c(0.1, 0.2), variance = 0.6)

# evaluate it on a greta array to get the variance-covariance matrix
x <- greta_array(rnorm(8), dim = c(4, 2))
k1(x)

# non-symmetric covariance between two sets of points
x2 <- greta_array(rnorm(10), dim = c(5, 2))
k1(x, x2)

# create a bias kernel, with the variance as a variable
k2 <- bias(variance = lognormal(0, 1))

# combine two kernels and evaluate
K <- k1 + k2
K(x, x2)
}
